fix end game modal
